chore: use pytest
