+++ packages/cli/tests/workerd-test/asgi/pyproject.toml
@@ -2,4 +2,4 @@
 name = "test"
 version = "0.0.0"
 requires-python = ">=3.12"
-dependencies = ["pytest"]
+dependencies = ["pytest", "pytest-asyncio<1.2.0"]
