+++ packages/cli/tests/workerd-test/asgi/worker.py
@@ -1,10 +1,26 @@
 import asyncio
-import logging
+import os
+import sys
 
 import asgi
-import js
-from pyodide.ffi import to_js
-from workers import Request, WorkerEntrypoint
+import pytest
+from pyodide.webloop import WebLoop
+from workers import WorkerEntrypoint
+
+
+async def noop(*args):
+    pass
+
+
+# pytest-asyncio relies on these but in Pyodide < 0.29 WebLoop does not implement them
+WebLoop.shutdown_asyncgens = noop
+WebLoop.shutdown_default_executor = noop
+
+# Pyodide 0.26.0a2's _cancel_all_tasks calls task.exception() on pending tasks,
+# which raises InvalidStateError under Pyodide's WebLoop.
+if sys.version_info < (3, 13):
+    asyncio.runners._cancel_all_tasks = lambda loop: None  # type: ignore[attr-defined]
+
 
 # ---------------------------------------------------------------------------
 # ASGI apps
@@ -183,524 +199,9 @@ class Default(WorkerEntrypoint):
         elif path == "/stream":
             return await asgi.fetch(streaming_app, request, self.env, self.ctx)
 
-        # Verify that `asgi` can handle JS-style headers and Python-style headers:
-        js_request = js.Request.new("http://example.com/", headers=to_js(example_hdr))
-        py_request = Request("http://example.com/", headers=example_hdr)
-        js_scope = asgi.request_to_scope(js_request, self.env)
-        py_scope = asgi.request_to_scope(py_request, self.env)
-        assert (
-            js_scope["headers"]
-            == py_scope["headers"]
-            == [(k.lower().encode(), v.encode()) for k, v in example_hdr.items()]
-        )
-
-        # Standard asgi.fetch test path.
         return await asgi.fetch(app, request, self.env, self.ctx)
 
-    async def test(self, ctrl):
-        await test_headers(self.env)
-        await test_sse(self.env)
-        await test_streaming(self.env)
-        await test_error_after_response_is_logged(self.env)
-        await test_background_task_error_is_logged()
-        await test_app_exception_before_response_is_logged()
-        await test_lifespan_unsupported_app_does_not_hang(self.env)
-        await test_lifespan_full_cycle(self.env)
-        await test_lifespan_startup_failed_propagates(self.env)
-        await test_lifespan_shutdown_failed_propagates(self.env)
-        await test_null_body_statuses(self.env)
-        await test_response_start_without_headers(self.env)
-        await test_response_body_without_body_key(self.env)
-
-
-# ---------------------------------------------------------------------------
-# Test: headers
-# ---------------------------------------------------------------------------
-
-
-async def test_headers(env):
-    response = await env.SELF.fetch("http://example.com/", headers=to_js(example_hdr))
-    for header in response.headers.items():
-        assert isinstance(header[0], str) and isinstance(header[1], str)
-        expected_hdr = {k.lower(): v.lower() for k, v in example_hdr.items()}
-        assert header[0] in expected_hdr.keys()
-        assert expected_hdr[header[0]] == header[1].lower()
-
-
-# ---------------------------------------------------------------------------
-# Test: SSE (Server-Sent Events)
-# ---------------------------------------------------------------------------
-
-
-async def test_sse(env):
-    response = await env.SELF.fetch("http://example.com/sse")
-
-    assert response.headers["content-type"] == "text/event-stream; charset=utf-8"
-    assert response.headers["cache-control"] == "no-store"
-
-    from js import TextDecoder
-
-    reader = response.body.getReader()
-    content = ""
-    decoder = TextDecoder.new()
-
-    while True:
-        result = await reader.read()
-        if result.done:
-            break
-        chunk_text = decoder.decode(result.value, {"stream": True})
-        content += chunk_text
-
-    # Final flush
-    content += decoder.decode()
-    # Verify the expected events are in the response
-    assert "event: endpoint" in content
-    assert "data: /messages/?session_id=test123" in content
-    assert ": ping - message 1" in content
-    assert ": ping - message 2" in content
-    assert ": ping - message 3" in content
-
-
-# ---------------------------------------------------------------------------
-# Test: non-SSE multi-chunk streaming
-# Reproduction test for bug https://github.com/cloudflare/workers-py/issues/67
-# ---------------------------------------------------------------------------
-
-
-async def test_streaming(env):
-    response = await env.SELF.fetch("http://example.com/stream")
-
-    assert response.status == 200
-    assert response.headers["content-type"] == "application/octet-stream"
-
-    # Read the full response body via ReadableStream (same pattern as SSE test)
-    reader = response.body.getReader()
-    body_bytes = b""
-
-    while True:
-        result = await reader.read()
-        if result.done:
-            break
-        body_bytes += result.value.to_bytes()
-
-    expected_size = STREAMING_CHUNK_SIZE * STREAMING_NUM_CHUNKS  # 5120 bytes
-
-    # This is the key assertion: with bug #10, only the first chunk (1024 bytes)
-    # is returned. The fix should deliver all 5120 bytes.
-    assert len(body_bytes) == expected_size, (
-        f"Expected {expected_size} bytes, got {len(body_bytes)}. "
-        f"Only {len(body_bytes) // STREAMING_CHUNK_SIZE} of {STREAMING_NUM_CHUNKS} chunks delivered."
-    )
-
-    # Verify each chunk has the correct content
-    for i in range(STREAMING_NUM_CHUNKS):
-        start = i * STREAMING_CHUNK_SIZE
-        end = start + STREAMING_CHUNK_SIZE
-        expected_byte = i % 256
-        chunk = body_bytes[start:end]
-        assert all(b == expected_byte for b in chunk), (
-            f"Chunk {i}: expected all bytes to be {expected_byte}, "
-            f"but got different content"
-        )
-
-
-# ---------------------------------------------------------------------------
-# Logging tests
-# ---------------------------------------------------------------------------
-
-
-class _ListHandler(logging.Handler):
-    """A logging handler that captures records into a list for assertions."""
-
-    def __init__(self):
-        super().__init__()
-        self.records: list[logging.LogRecord] = []
-
-    def emit(self, record):
-        self.records.append(record)
-
-    def clear(self):
-        self.records.clear()
-
-
-def _install_handler():
-    """Install a ListHandler on the 'asgi' logger and return it."""
-    handler = _ListHandler()
-    logger = logging.getLogger("asgi")
-    logger.addHandler(handler)
-    # Ensure the logger level is low enough to capture everything.
-    logger.setLevel(logging.DEBUG)
-    return handler
-
-
-def _remove_handler(handler):
-    logging.getLogger("asgi").removeHandler(handler)
-
-
-# -- Test: error after response already sent is logged (not swallowed) --------
-
-
-class _ErrorAfterResponseApp:
-    """ASGI app that sends a valid response, then raises an exception."""
-
-    async def __call__(self, scope, receive, send):
-        if scope["type"] == "lifespan":
-            message = await receive()
-            if message["type"] == "lifespan.startup":
-                await send({"type": "lifespan.startup.complete"})
-            return
-
-        if scope["type"] == "http":
-            await receive()
-            await send(
-                {
-                    "type": "http.response.start",
-                    "status": 200,
-                    "headers": [(b"content-type", b"text/plain")],
-                }
-            )
-            await send(
-                {
-                    "type": "http.response.body",
-                    "body": b"ok",
-                }
-            )
-            # Response is already sent — now raise an error.
-            raise RuntimeError("post-response error for testing")
-
-
-async def test_error_after_response_is_logged(env):
-    handler = _install_handler()
-    try:
-        req = js.Request.new("http://example.com/log-test")
-        # The response should still succeed — the error happens after it's sent.
-        response = await asgi.fetch(_ErrorAfterResponseApp(), req, env)
-        assert response.status == 200, f"Expected 200, got {response.status}"
-
-        # consume the body
-        await response.arrayBuffer()
-
-        # Let the event loop run
-        await asyncio.sleep(0.5)
-
-        # The error should have been logged, not swallowed.
-        error_records = [r for r in handler.records if r.levelno >= logging.ERROR]
-        assert len(error_records) > 0, (
-            "Expected at least one ERROR log record for post-response exception, "
-            f"got {len(error_records)}. All records: {[r.getMessage() for r in handler.records]}"
-        )
-        matched = any(
-            "post-response error for testing" in str(r.exc_info[1])
-            if r.exc_info
-            else False
-            for r in error_records
-        )
-        assert matched, (
-            "Expected log message containing 'post-response error for testing', "
-            f"got: {[r.getMessage() for r in error_records]}"
-        )
-    finally:
-        _remove_handler(handler)
-
-
-# -- Test: background task error is logged ------------------------------------
-
-
-async def test_background_task_error_is_logged():
-    handler = _install_handler()
-    try:
-
-        async def failing_task():
-            raise ValueError("background task failure for testing")
-
-        asgi.run_in_background(failing_task())
-
-        # Let the event loop run
-        await asyncio.sleep(0.5)
-
-        error_records = [r for r in handler.records if r.levelno >= logging.ERROR]
-        assert len(error_records) > 0, (
-            "Expected at least one ERROR log for background task failure, "
-            f"got {len(error_records)}. All records: {[r.getMessage() for r in handler.records]}"
-        )
-        matched = any(
-            "background task failure for testing" in str(r.exc_info[1])
-            if r.exc_info
-            else False
-            for r in error_records
-        )
-        assert matched, (
-            "Expected log message containing 'background task failure for testing', "
-            f"got: {[r.getMessage() for r in error_records]}"
-        )
-    finally:
-        _remove_handler(handler)
-
-
-# -- Test: app exception before response is logged and re-raised -------------
-
-
-class _ErrorBeforeResponseApp:
-    """ASGI app that raises before sending any response."""
-
-    async def __call__(self, scope, receive, send):
-        if scope["type"] == "lifespan":
-            message = await receive()
-            if message["type"] == "lifespan.startup":
-                await send({"type": "lifespan.startup.complete"})
-            return
-
-        if scope["type"] == "http":
-            await receive()
-            raise RuntimeError("app crash before response for testing")
-
-
-async def test_app_exception_before_response_is_logged():
-    handler = _install_handler()
-    try:
-        req = js.Request.new("http://example.com/crash-test")
-        threw = False
-        try:
-            await asgi.fetch(_ErrorBeforeResponseApp(), req, {})
-        except RuntimeError as e:
-            threw = True
-            assert "app crash before response for testing" in str(e), (
-                f"Expected original exception message, got: {e}"
-            )
-
-        assert threw, "Expected RuntimeError to be raised from asgi.fetch"
-
-        # fetch() should have logged the error before re-raising.
-        error_records = [r for r in handler.records if r.levelno >= logging.ERROR]
-        assert len(error_records) > 0, (
-            "Expected at least one ERROR log for request failure, "
-            f"got {len(error_records)}. All records: {[r.getMessage() for r in handler.records]}"
-        )
-        matched = any("ASGI request failed" in r.getMessage() for r in error_records)
-        assert matched, (
-            "Expected log message containing 'ASGI request failed', "
-            f"got: {[r.getMessage() for r in error_records]}"
-        )
-    finally:
-        _remove_handler(handler)
-
-
-# ---------------------------------------------------------------------------
-# Test: lifespan protocol tolerance (app that rejects the lifespan scope)
-# ---------------------------------------------------------------------------
-
-
-class _NoLifespanApp:
-    """Mimics Django's ASGIHandler, which raises on any non-http scope."""
-
-    async def __call__(self, scope, receive, send):
-        if scope["type"] != "http":
-            raise ValueError(f"can only handle http, not {scope['type']}")
-        await receive()
-        await send(
-            {
-                "type": "http.response.start",
-                "status": 200,
-                "headers": [(b"content-type", b"text/plain")],
-            }
-        )
-        await send({"type": "http.response.body", "body": b"served"})
-
-
-async def test_lifespan_unsupported_app_does_not_hang(env):
-    req = js.Request.new("http://example.com/no-lifespan")
-    # The previous adaptor blocked forever waiting for a startup ack the app
-    # never sends; wait_for turns that hang into an observable test failure.
-    response = await asyncio.wait_for(asgi.fetch(_NoLifespanApp(), req, env), timeout=5)
-    assert response.status == 200
-    assert await response.text() == "served"
-
-
-# ---------------------------------------------------------------------------
-# Test: null-body statuses must not carry a response body
-# ---------------------------------------------------------------------------
-
-
-class _NullBodyApp:
-    def __init__(self, status):
-        self._status = status
-
-    async def __call__(self, scope, receive, send):
-        if scope["type"] == "lifespan":
-            message = await receive()
-            if message["type"] == "lifespan.startup":
-                await send({"type": "lifespan.startup.complete"})
-            return
-        if scope["type"] != "http":
-            return
-        await receive()
-        await send(
-            {
-                "type": "http.response.start",
-                "status": self._status,
-                "headers": [(b"x-null-body", b"1")],
-            }
-        )
-        await send({"type": "http.response.body", "body": b"body-should-be-dropped"})
-
-
-async def test_null_body_statuses(env):
-    for status in (204, 205, 304):
-        req = js.Request.new(f"http://example.com/null-body/{status}")
-        response = await asgi.fetch(_NullBodyApp(status), req, env)
-        assert response.status == status, f"expected {status}, got {response.status}"
-        assert response.headers["x-null-body"] == "1"
-        assert await response.text() == ""
-
-
-# ---------------------------------------------------------------------------
-# Test: defensive handling of ASGI messages missing optional keys
-# ---------------------------------------------------------------------------
-
-
-class _MissingHeadersApp:
-    async def __call__(self, scope, receive, send):
-        if scope["type"] == "lifespan":
-            message = await receive()
-            if message["type"] == "lifespan.startup":
-                await send({"type": "lifespan.startup.complete"})
-            return
-        if scope["type"] != "http":
-            return
-        await receive()
-        await send({"type": "http.response.start", "status": 200})
-        await send({"type": "http.response.body", "body": b"ok"})
-
-
-async def test_response_start_without_headers(env):
-    req = js.Request.new("http://example.com/missing-headers")
-    response = await asgi.fetch(_MissingHeadersApp(), req, env)
-    assert response.status == 200
-    assert await response.text() == "ok"
-
-
-class _MissingBodyKeyApp:
-    async def __call__(self, scope, receive, send):
-        if scope["type"] == "lifespan":
-            message = await receive()
-            if message["type"] == "lifespan.startup":
-                await send({"type": "lifespan.startup.complete"})
-            return
-        if scope["type"] != "http":
-            return
-        await receive()
-        await send(
-            {
-                "type": "http.response.start",
-                "status": 200,
-                "headers": [(b"content-type", b"text/plain")],
-            }
-        )
-        await send({"type": "http.response.body"})
-
-
-async def test_response_body_without_body_key(env):
-    req = js.Request.new("http://example.com/missing-body")
-    response = await asgi.fetch(_MissingBodyKeyApp(), req, env)
-    assert response.status == 200
-    assert await response.text() == ""
-
-
-# ---------------------------------------------------------------------------
-# Test: full lifespan cycle (startup AND shutdown are both delivered)
-# ---------------------------------------------------------------------------
-
-
-class _LifespanCycleApp:
-    def __init__(self):
-        self.events = []
-
-    async def __call__(self, scope, receive, send):
-        if scope["type"] == "lifespan":
-            while True:
-                message = await receive()
-                if message["type"] == "lifespan.startup":
-                    self.events.append("startup")
-                    await send({"type": "lifespan.startup.complete"})
-                elif message["type"] == "lifespan.shutdown":
-                    self.events.append("shutdown")
-                    await send({"type": "lifespan.shutdown.complete"})
-                    return
-        if scope["type"] != "http":
-            return
-        await receive()
-        await send(
-            {
-                "type": "http.response.start",
-                "status": 200,
-                "headers": [(b"content-type", b"text/plain")],
-            }
-        )
-        await send({"type": "http.response.body", "body": b"ok"})
-
-
-async def test_lifespan_full_cycle(env):
-    app = _LifespanCycleApp()
-    req = js.Request.new("http://example.com/lifespan-cycle")
-    response = await asgi.fetch(app, req, env)
-    assert response.status == 200
-    assert app.events == ["startup", "shutdown"]
-
-
-# ---------------------------------------------------------------------------
-# Test: lifespan startup/shutdown failures propagate out of asgi.fetch
-# ---------------------------------------------------------------------------
-
-
-class _StartupFailApp:
-    async def __call__(self, scope, receive, send):
-        if scope["type"] == "lifespan":
-            await receive()
-            await send({"type": "lifespan.startup.failed", "message": "boom-startup"})
-            return
-        if scope["type"] != "http":
-            return
-        await receive()
-        await send({"type": "http.response.start", "status": 200, "headers": []})
-        await send({"type": "http.response.body", "body": b"ok"})
-
-
-async def test_lifespan_startup_failed_propagates(env):
-    req = js.Request.new("http://example.com/startup-fail")
-    raised = None
-    try:
-        await asyncio.wait_for(asgi.fetch(_StartupFailApp(), req, env), timeout=5)
-    except Exception as e:
-        raised = e
-    assert isinstance(raised, RuntimeError), f"expected RuntimeError, got {raised!r}"
-    assert "boom-startup" in str(raised)
-
-
-class _ShutdownFailApp:
-    async def __call__(self, scope, receive, send):
-        if scope["type"] == "lifespan":
-            while True:
-                message = await receive()
-                if message["type"] == "lifespan.startup":
-                    await send({"type": "lifespan.startup.complete"})
-                elif message["type"] == "lifespan.shutdown":
-                    await send(
-                        {"type": "lifespan.shutdown.failed", "message": "boom-shutdown"}
-                    )
-                    return
-        if scope["type"] != "http":
-            return
-        await receive()
-        await send({"type": "http.response.start", "status": 200, "headers": []})
-        await send({"type": "http.response.body", "body": b"ok"})
-
-
-async def test_lifespan_shutdown_failed_propagates(env):
-    req = js.Request.new("http://example.com/shutdown-fail")
-    raised = None
-    try:
-        await asyncio.wait_for(asgi.fetch(_ShutdownFailApp(), req, env), timeout=5)
-    except Exception as e:
-        raised = e
-    assert isinstance(raised, RuntimeError), f"expected RuntimeError, got {raised!r}"
-    assert "boom-shutdown" in str(raised)
+    async def test(self):
+        os.chdir("/session/metadata/tests")
+        args = [".", "-vv"]
+        assert pytest.main(args) == 0
